# tests/periph_rtc: test sub-second precision