<!--
The RIOT community cares a lot about code quality.
Therefore, before describing what your contribution is about, we would like
you to make sure that your modifications are compliant with the RIOT
coding conventions, see https://github.com/RIOT-OS/RIOT/wiki/Coding-conventions.
-->

### Contribution description
This expands `tests/periph_rtc` to verify the microsecond precision of RTC timers. This is needed to test that an RTC alarm fires at the desired time and not with an arbitrary offset up to +/-1 second. It also tests that `rtc_set_time()` behaves equivalently correctly, which is what I needed to test when I wrote this.

<!--
Put here the description of your contribution:
- describe which part(s) of RIOT is (are) involved
- if it's a bug fix, describe the bug that it solves and how it is solved
- you can also give more information to reviewers about how to test your changes
-->


### Testing procedure
Run `tests/periph_rtc` and observe new output (`error -177429 microseconds`).

Currently it fails for `kinetis/kw41z` (perhaps all `kinetis`) and that looks like this:
```
main(): This is RIOT! (Version: 2018.10-RC1-679-ga296b-telnet-shell)

RIOT RTC low-level driver test
This test will display 'Alarm!' every 2 seconds for 4 times
  Setting clock to   2011-12-13 14:15:15
Clock value is now   2011-12-13 14:15:15
  Setting alarm to   2011-12-13 14:15:17
   Alarm is set to   2011-12-13 14:15:17

[30303] First alarm expected at 2018615 microseconds
[1841186] Alarm! after 1822571 microseconds (error -177429 microseconds)
[3841186] Alarm! after 3822571 microseconds (error -177429 microseconds)
[5841186] Alarm! after 5822571 microseconds (error -177429 microseconds)
[7841186] Alarm! after 7822571 microseconds (error -177429 microseconds)
```
In the case of `kw41z` the error is always a random value (abs(error) < 1 second) depending on what value is randomly in the RTC subsecond register which is currently ignored by `kinetis/periph/rtt_set_counter()`.

A successful test should have output similar to:
```
main(): This is RIOT! (Version: 2018.10-RC1-680-g83348-telnet-shell)

RIOT RTC low-level driver test
This test will display 'Alarm!' every 2 seconds for 4 times
  Setting clock to   2011-12-13 14:15:15
Clock value is now   2011-12-13 14:15:15
  Setting alarm to   2011-12-13 14:15:17
   Alarm is set to   2011-12-13 14:15:17

[30303] First alarm expected at 2018585 microseconds
[2018615] Alarm! after 2000030 microseconds (error 30 microseconds)
[4018615] Alarm! after 4000030 microseconds (error 30 microseconds)
[6018615] Alarm! after 6000030 microseconds (error 30 microseconds)
[8018615] Alarm! after 8000030 microseconds (error 30 microseconds)
```
The error precision is limited by the frequency of `xtimer`'s timer, hence the 30us error for `kw41z` with a 16bit 32kHz `xtimer` where 30us is only 1 clock tick. Platforms with a faster `xtimer` should have a smaller error. Perhaps no platform should have an error over 100us.

<!--
Details steps to test your contribution:
- which test/example to compile for which board and is there a 'test' command
- how to know that it was not working/available in master
- the expected success test output
-->

### Comments
Must test other platforms.
Perhaps I should have updated the readme.
Why doesn't newlib declare `settimeofday()`?

### Issues/PRs references
#10764 fixes `kinetis/periph_rtt` to give the successful test output shown above.

<!--
Examples: Fixes #1234. See also #5678. Depends on PR #9876.

Please use keywords (e.g., fixes, resolve) with the links to the issues you
resolved, this way they will be automatically closed when your pull request
is merged. See https://help.github.com/articles/closing-issues-using-keywords/.
-->