@benemorius 2019-01-14
> I think this test is useful and necessary, although I'm concerned about the case where the RTC and Xtimer are powered by the same clock. Wouldn't the test stop working in that case?

It works for me on kw41z where both xtimer and periph_rtc use the same low power clock source. I think most low power platforms will be this way. I can't think of a reason it should be a problem.