@PeterKietzmann 2019-01-23
I think this change is useful. It already helped in #10764.

> I'm concerned about the case where the RTC and Xtimer are powered by the same clock.

@jcarrano I'm no convinced about that but also lacking knowledge about xtimers internals so I'd rather run some tests on exemplary devices. Do you know of any where your concerns apply?