@aabadie 2019-01-23
> I think this change is useful.

I fully agree.
The Python test automation script should be updated, otherwise it's broken in the current state.