@benemorius 2019-01-23
> The Python test automation script should be updated, otherwise it's broken in the current state.

I didn't see those scripts until now. Wow that's a very nice system.

Should we check that the error is less than some predetermined value? I didn't expect to see any platform with an error over 100us but @PeterKietzmann found in #10764 that it can be higher. Although it's possible that that's due to another problem.