@PeterKietzmann 2019-01-24
> Should we check that the error is less than some predetermined value?

I think this would be a nice improvement. However, this requires certain board specific assumptions about maximum deviations. @MrKevinWeiss came to the great idea to look up worst-case parameters of clock sources at DigiKey which lead to a total number around 500ppm. With that we would at least find misconfigured timers