@PeterKietzmann 2019-01-24
BTW: @MrKevinWeiss would you like to take this PR over?