@MrKevinWeiss 2019-01-24
Only for +-250ppm for frequency stability and +-250ppm for frequency tolerance (not necessarily compounded but we can say that as a dumb worst case) for crystals...  Not to mention internal RC oscillators.
@PeterKietzmann Like is a strong word, but if we need it to be finished and added in to be runnable with the compile_and_test_for_board.py tool, then I will, it will only be after the uart test PR is merged though, FIFO style.