@benpicco 2020-02-26
That's pretty useful (also for RTT) - want to give it a rebase?