@MrKevinWeiss 2020-02-26
I have #13483 to replace this.  It takes care of the conflict and adapts the python test.